Move utility drawable-to-bitmap logic to separate file.
