+++ app/src/main/java/lt/gyvosistorijos/MainController.kt
@@ -1,9 +1,5 @@
 package lt.gyvosistorijos
 
-import android.graphics.Bitmap
-import android.graphics.Canvas
-import android.graphics.drawable.BitmapDrawable
-import android.graphics.drawable.Drawable
 import android.location.Location
 import android.support.v4.content.ContextCompat
 import android.view.LayoutInflater
@@ -24,6 +20,7 @@ import lt.gyvosistorijos.location.LocationService
 import lt.gyvosistorijos.manager.RemoteConfigManager
 import lt.gyvosistorijos.utils.AppEvent
 import lt.gyvosistorijos.utils.distanceMetersTo
+import lt.gyvosistorijos.utils.drawableToBitmap
 import timber.log.Timber
 
 class MainController : Controller(), LocationListener, GoogleMap.OnMarkerClickListener {
@@ -200,26 +197,4 @@ class MainController : Controller(), LocationListener, GoogleMap.OnMarkerClickLi
 
         (activity as MainActivity).geofenceHelper.setGeofenceRegions(geofenceRegions)
     }
-}
-
-private fun drawableToBitmap(drawable: Drawable): Bitmap {
-    if (drawable is BitmapDrawable) {
-        if (drawable.bitmap != null) {
-            return drawable.bitmap
-        }
-    }
-
-    val bitmap: Bitmap
-    if (drawable.intrinsicWidth <= 0 || drawable.intrinsicHeight <= 0) {
-        bitmap = Bitmap.createBitmap(1, 1,
-                Bitmap.Config.ARGB_8888) // Single color bitmap will be created of 1x1 pixel
-    } else {
-        bitmap = Bitmap.createBitmap(drawable.intrinsicWidth, drawable.intrinsicHeight,
-                Bitmap.Config.ARGB_8888)
-    }
-
-    val canvas = Canvas(bitmap)
-    drawable.setBounds(0, 0, canvas.width, canvas.height)
-    drawable.draw(canvas)
-    return bitmap
 }
