+++ app/src/main/java/lt/gyvosistorijos/utils/DrawableUtils.kt
@@ -1,0 +1,28 @@
+package lt.gyvosistorijos.utils
+
+import android.graphics.Bitmap
+import android.graphics.Canvas
+import android.graphics.drawable.BitmapDrawable
+import android.graphics.drawable.Drawable
+
+fun drawableToBitmap(drawable: Drawable): Bitmap {
+    if (drawable is BitmapDrawable) {
+        if (drawable.bitmap != null) {
+            return drawable.bitmap
+        }
+    }
+
+    val bitmap: Bitmap
+    if (drawable.intrinsicWidth <= 0 || drawable.intrinsicHeight <= 0) {
+        bitmap = Bitmap.createBitmap(1, 1,
+                Bitmap.Config.ARGB_8888) // Single color bitmap will be created of 1x1 pixel
+    } else {
+        bitmap = Bitmap.createBitmap(drawable.intrinsicWidth, drawable.intrinsicHeight,
+                Bitmap.Config.ARGB_8888)
+    }
+
+    val canvas = Canvas(bitmap)
+    drawable.setBounds(0, 0, canvas.width, canvas.height)
+    drawable.draw(canvas)
+    return bitmap
+}
